Make `generate-javadoc-jar` trigger on CI env.
the `ci-build` profile triggers on `env.SCM_REVISION` but yet `generate-javadoc-jar` was auto-triggering based on `env.JOB_NAME`; this is inconsistent. these profiles could be combined into one (`ci-build`) but that would make it harder to generate Javadocs without having to publish a full release. the compromise is to keep both profiles but make sure they trigger on the same environment vars.